hide implementation from RGW
